ml-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pdfminer.six==20231228
python-docx==1.1.2
